Rename de Package
-Versão do App: 0.4.12 -> 0.4.13;

-Package retrofitServerService alterado para api e refatorado todas as classes referêntes.
