+++ app/src/main/java/br/com/garrav/projetogarrav/api/EventServerService.java
@@ -1,4 +1,4 @@
-package br.com.garrav.projetogarrav.retrofitServerService;
+package br.com.garrav.projetogarrav.api;
 
 import android.app.ProgressDialog;
 import android.content.Context;
